First part of info endpoint completed - tags

+++ src/main/java/czdbdk/dbdkbe/models/parameters/BookSizeInfo.java
@@ -14,12 +14,12 @@ public class BookSizeInfo {
     private String slug;
     private String name;
     private Long booksMatchesValue;
-    private Long minPages;
-    private Long maxPages;
+    private int minPages;
+    private int maxPages;
     @Autowired
     BookRepository bookRepository;
 
-    public BookSizeInfo(String name, Long minPages, Long maxPages){
+    public BookSizeInfo(String name, int minPages, int maxPages){
         this.name = name;
         this.minPages = minPages;
         this.maxPages = maxPages;
